back up

+++ codec/encoder/core/inc/rc.h
@@ -175,13 +175,13 @@ int32_t   iBitRate;
 int32_t   iPreviousBitrate;
 int32_t   iPreviousGopSize;
 double    fFrameRate;
-int64_t   iBitsPerFrame;
-int64_t   iMaxBitsPerFrame;
+int32_t   iBitsPerFrame;
+int32_t   iMaxBitsPerFrame;
 double    dPreviousFps;
 
 // bits allocation and status
-int64_t   iRemainingBits;
-int64_t   iTargetBits;
+int32_t   iRemainingBits;
+int32_t   iTargetBits;
 int32_t   iCurrentBitsLevel;//0:normal; 1:limited; 2:exceeded.
 
 int32_t   iIdrNum;
@@ -226,7 +226,7 @@ int32_t   iLastCalculatedQScale;
 
 //for skip frame and padding
 int32_t   iBufferSizeSkip;
-int64_t   iBufferFullnessSkip;
+int32_t   iBufferFullnessSkip;
 int32_t   iBufferMaxBRFullness[TIME_WINDOW_TOTAL];//0: EVEN_TIME_WINDOW; 1: ODD_TIME_WINDOW
 int32_t   iPredFrameBit;
 bool      bNeedShiftWindowCheck[TIME_WINDOW_TOTAL];
